web/requirements.txt
asgiref>=3.2.7
certifi>=2020.6.20
chardet>=3.0.4
dj-static>=0.0.6
Django>=3.1.1
djangorestframework>=3.11.1
Pillow>=7.2.0
psycopg2-binary>=2.8.5
pytz>=2019.3
requests>=2.25.1
sqlparse>=0.3.1
static3>=0.7.0
pycryptodome>=3.9.8
urllib3>=1.25.10
dj-database-url>=0.5.0
httpx>=0.16.1
model_mommy>=2.0.0
django-cors-headers>=3.2.1
django-rest-knox>=4.1.0
django_crispy_forms>=1.11